Move RestrictAccess and JWTParse to Security namespace, add const role strings, update UserInformationModel
