+++ BridgeCareApp/BridgeCare/Controllers/AttributeController.cs
@@ -1,4 +1,5 @@
 ﻿using BridgeCare.Interfaces;
+using BridgeCare.Security;
 using System;
 using System.Web.Http;
 
@@ -21,7 +22,7 @@ namespace BridgeCare.Controllers
         /// <returns>IHttpActionResult</returns>
         [HttpGet]
         [Route("api/GetAttributes")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult GetAttributes() => Ok(repo.GetAttributes(db));
     }
 }
