+++ BridgeCareApp/BridgeCare/Controllers/CriteriaDrivenBudgetsController.cs
@@ -1,5 +1,6 @@
 ﻿using BridgeCare.Interfaces.CriteriaDrivenBudgets;
 using BridgeCare.Models.CriteriaDrivenBudgets;
+using BridgeCare.Security;
 using System;
 using System.Collections.Generic;
 using System.Web.Http;
@@ -27,7 +28,7 @@ namespace BridgeCare.Controllers
         [HttpGet]
         [Route("api/GetCriteriaDrivenBudgets/{id}")]
         [ModelValidation("The scenario id is invalid.")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult GetCriteriaDrivenBudgets(int id) => Ok(repo.GetCriteriaDrivenBudgets(id, db));
 
         /// <summary>
@@ -39,7 +40,7 @@ namespace BridgeCare.Controllers
         [HttpPost]
         [Route("api/SaveCriteriaDrivenBudgets/{id}")]
         [ModelValidation("The criteria driven budgets data is invalid.")]
-        [Filters.RestrictAccess("PD-BAMS-Administrator", "PD-BAMS-DBEngineer")]
+        [RestrictAccess(Role.ADMINISTRATOR, Role.DISTRICT_ENGINEER)]
         public IHttpActionResult SaveCriteriaDrivenBudgets(int id, [FromBody]List<CriteriaDrivenBudgetsModel> models)
         {
             var result = repo.SaveCriteriaDrivenBudgets(id, models, db);
