+++ BridgeCareApp/BridgeCare/Controllers/DeficientController.cs
@@ -3,6 +3,7 @@ using System.Web.Http;
 using System.Web.Http.Filters;
 using BridgeCare.Interfaces;
 using BridgeCare.Models;
+using BridgeCare.Security;
 
 namespace BridgeCare.Controllers
 {
@@ -27,7 +28,7 @@ namespace BridgeCare.Controllers
         [HttpGet]
         [Route("api/GetScenarioDeficientLibrary/{id}")]
         [ModelValidation("The scenario id is invalid.")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult GetSimulationDeficientLibrary(int id)
             => Ok(repo.GetSimulationDeficientLibrary(id, db));
 
@@ -39,7 +40,7 @@ namespace BridgeCare.Controllers
         [HttpPost]
         [Route("api/SaveScenarioDeficientLibrary")]
         [ModelValidation("The deficient data is invalid.")]
-        [Filters.RestrictAccess("PD-BAMS-Administrator", "PD-BAMS-DBEngineer")]
+        [RestrictAccess(Role.ADMINISTRATOR, Role.DISTRICT_ENGINEER)]
         public IHttpActionResult SaveSimulationDeficientLibrary([FromBody] DeficientLibraryModel model)
             => Ok(repo.SaveSimulationDeficientLibrary(model, db));
     }
