+++ BridgeCareApp/BridgeCare/Controllers/InvestmentLibraryController.cs
@@ -1,5 +1,6 @@
 ﻿using BridgeCare.Interfaces;
 using BridgeCare.Models;
+using BridgeCare.Security;
 using System;
 using System.Web.Http;
 using System.Web.Http.Filters;
@@ -29,7 +30,7 @@ namespace BridgeCare.Controllers
         [HttpGet]
         [Route("api/GetScenarioInvestmentLibrary/{id}")]
         [ModelValidation("The scenario id is invalid.")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult GetSimulationInvestmentLibrary(int id)
              => Ok(repo.GetSimulationInvestmentLibrary(id, db));
 
@@ -41,7 +42,7 @@ namespace BridgeCare.Controllers
         [HttpPost]
         [Route("api/SaveScenarioInvestmentLibrary")]
         [ModelValidation("Given investment data is not valid")]
-        [Filters.RestrictAccess("PD-BAMS-Administrator", "PD-BAMS-DBEngineer")]
+        [RestrictAccess(Role.ADMINISTRATOR, Role.DISTRICT_ENGINEER)]
         public IHttpActionResult SaveSimulationInvestmentLibrary([FromBody]InvestmentLibraryModel model)
             => Ok(repo.SaveSimulationInvestmentLibrary(model, db));
     }
