+++ BridgeCareApp/BridgeCare/Controllers/NetworksController.cs
@@ -1,4 +1,5 @@
 ﻿using BridgeCare.Interfaces;
+using BridgeCare.Security;
 using System;
 using System.Web.Http;
 
@@ -21,7 +22,7 @@ namespace BridgeCare.Controllers
         /// <returns>IHttpActionResult</returns>
         [HttpGet]
         [Route("api/GetNetworks")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult GetNetworks() => Ok(repo.GetAllNetworks(db));
     }
 }
