+++ BridgeCareApp/BridgeCare/Controllers/TargetController.cs
@@ -3,6 +3,7 @@ using System.Web.Http;
 using System.Web.Http.Filters;
 using BridgeCare.Interfaces;
 using BridgeCare.Models;
+using BridgeCare.Security;
 
 namespace BridgeCare.Controllers
 {
@@ -27,7 +28,7 @@ namespace BridgeCare.Controllers
         [HttpGet]
         [Route("api/GetScenarioTargetLibrary/{id}")]
         [ModelValidation("The scenario id is invalid.")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult GetSimulationTargetLibrary(int id) =>
             Ok(repo.GetSimulationTargetLibrary(id, db));
 
@@ -39,7 +40,7 @@ namespace BridgeCare.Controllers
         [HttpPost]
         [Route("api/SaveScenarioTargetLibrary")]
         [ModelValidation("The target data is invalid.")]
-        [Filters.RestrictAccess("PD-BAMS-Administrator", "PD-BAMS-DBEngineer")]
+        [RestrictAccess(Role.ADMINISTRATOR, Role.DISTRICT_ENGINEER)]
         public IHttpActionResult SaveSimulationTargetLibrary([FromBody]TargetLibraryModel model) => 
             Ok(repo.SaveSimulationTargetLibrary(model, db));
     }
