+++ BridgeCareApp/BridgeCare/Controllers/ValidationController.cs
@@ -1,5 +1,6 @@
 ﻿using BridgeCare.Interfaces;
 using BridgeCare.Models;
+using BridgeCare.Security;
 using System;
 using System.Web.Http;
 using System.Web.Http.Filters;
@@ -25,7 +26,7 @@ namespace BridgeCare.Controllers
         [HttpPost]
         [Route("api/ValidateEquation")]
         [ModelValidation("The equation data is invalid.")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult ValidateEquation(ValidateEquationModel model)
         {
             repo.ValidateEquation(model, db);
@@ -40,7 +41,7 @@ namespace BridgeCare.Controllers
         [HttpPost]
         [Route("api/ValidateCriteria")]
         [ModelValidation("The criteria data is invalid.")]
-        [Filters.RestrictAccess]
+        [RestrictAccess]
         public IHttpActionResult ValidateCriteria([FromBody]ValidateCriteriaModel model) =>
             Ok(repo.ValidateCriteria(model.Criteria, db));
     }
