+++ BridgeCareApp/BridgeCare/Security/Role.cs
@@ -1,0 +1,10 @@
+﻿namespace BridgeCare.Security
+{
+    public static class Role
+    {
+        public const string ADMINISTRATOR = "PD-BAMS-Administrator";
+        public const string DISTRICT_ENGINEER = "PD-BAMS-DBEngineer";
+        public const string PLANNING_PARTNER = "PD-BAMS-PlanningPartner";
+        public const string CWOPA = "pd-BAMS-CWOPA";
+    }
+}
